move main 'lib/api/index' to 'lib/index' - rename all partial api scripts - update require pathing
