+++ lib/api/clear.js
