+++ lib/api/parallel.js
