+++ lib/api/serial.js
