+++ lib/cli/spawn.js
@@ -3,7 +3,7 @@
 const Promise = require('bluebird');
 const find = require('../utils/find');
 const loadPlugins = require('../plugins').load
-const Fly = require('../api');
+const Fly = require('../');
 const co = Promise.coroutine;
 
 /**
