+++ lib/index.js
@@ -1,7 +1,8 @@
 'use strict';
 
-const $ = require('../utils');
-const Emitter = require('../emitter');
+const $ = require('./utils');
+const API = require('./api');
+const Emitter = require('./emitter');
 
 class Fly extends Emitter {
 	constructor(opts) {
@@ -19,13 +20,13 @@ class Fly extends Emitter {
 
 // attach methods to Fly.prototype
 Object.defineProperties(Fly.prototype, {
-	init: {value: require('./-init')},
-	clear: {value: require('./-clear')},
-	start: {value: require('./-start')},
-	serial: {value: require('./-serial')},
-	parallel: {value: require('./-parallel')},
-	source: {value: require('./-source')},
-	target: {value: require('./-target')}
+	init: {value: require('./api/init')},
+	clear: {value: require('./api/clear')},
+	start: {value: require('./api/start')},
+	serial: {value: require('./api/serial')},
+	parallel: {value: require('./api/parallel')},
+	source: {value: require('./api/source')},
+	target: {value: require('./api/target')}
 });
 
 module.exports = Fly;
